+++ src/Users/Notifications/LoginAddressNotification.php
@@ -2,14 +2,25 @@
 
 namespace Enraiged\Users\Notifications;
 
+use Enraiged\NetworkAddresses\Models\IpAddress;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Notifications\Messages\MailMessage;
 use Illuminate\Notifications\Notification;
 
-class LoginChangeNotification extends Notification implements ShouldQueue
+class LoginAddressNotification extends Notification implements ShouldQueue
 {
     use Queueable;
+
+    /**
+     *  Create an instance of the LoginAddressNotification.
+     *
+     *  @param  \Enraiged\NetworkAddresses\Models\IpAddress  $ipAddress
+     */
+    public function __construct(
+        private IpAddress $ipAddress,
+    )
+    {}
 
     /**
      *  Get the mail representation of the notification.
@@ -20,7 +31,7 @@ class LoginChangeNotification extends Notification implements ShouldQueue
     public function toMail($notifiable)
     {
         $message = (new MailMessage)
-            ->subject(__('Login Change'));
+            ->subject(__('New Device Login'));
 
         if ($notifiable->usernameIsEmailAddress) {
             $message->cc($notifiable->username, $notifiable->name);
@@ -28,13 +39,18 @@ class LoginChangeNotification extends Notification implements ShouldQueue
 
         if (config('enraiged.app.mail_markdown') === true) {
             return $message
-                ->markdown('mail.auth.login-change', ['user' => $notifiable]);
+                ->markdown('mail.auth.login-address', [
+                    'address' => $this->ipAddress,
+                    'user' => $notifiable,
+                ]);
         }
 
         return $message
             ->greeting(__('Hello :name', ['name' => $notifiable->name]))
-            ->line(__('There have been changes to your account login.'))
-            ->line(__('If this update is unexpected, please take immediate action to protect your account by changing your password.'))
+            ->line(__('There has just been a login to your account from a new device or ip address.'))
+            ->line(__('The ip address is: :address', ['address' => $this->ipAddress->ip_address]))
+            ->line(__('The login credential is: :credential', ['credential' => $this->ipAddress->credential]))
+            ->line(__('If this login is unexpected, please take immediate action to protect your account by changing your password.'))
             ->line(__('Feel free to contact the support team if you have questions or require assistance.'));
     }