+++ src/Users/Notifications/VerifySecondaryNotification.php
@@ -60,4 +60,15 @@ class VerifySecondaryNotification extends Notification implements ShouldQueue
             ]
         );
     }
+
+    /**
+     *  Get the notification's delivery channels.
+     *
+     *  @param  mixed  $notifiable
+     *  @return array
+     */
+    public function via($notifiable)
+    {
+        return $notifiable->notificationChannels();
+    }
 }