+++ src/Users/Notifications/WelcomeNotification.php
@@ -66,6 +66,6 @@ class WelcomeNotification extends Notification implements ShouldQueue
      */
     public function via($notifiable)
     {
-        return ['mail'];
+        return $notifiable->notificationChannels();
     }
 }